feat(ui-utils): formatDate util was created
