+++ packages/ui-utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from './cn';
 export * from './formatCurrencyValue';
+export * from './formatDate';
